requirements.txt
amqp==5.0.2
asgiref==3.2.10
billiard==3.6.3.0
celery==5.0.2
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.2
django-accounts==0.1
django-filter==2.4.0
django-grappelli==2.14.2
django-import-export==2.4.0
django-redis==4.12.1
django-rename-app==0.1.2
et-xmlfile==1.0.1
getmac==0.8.2
idna==2.10
importlib-metadata==2.0.0
jdcal==1.4.1
kombu==5.0.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.5
pkg-resources==0.0.0
prompt-toolkit==3.0.8
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.0
